+++ nbproject/project.properties
@@ -78,7 +78,7 @@ javafx.enable.concurrent.external.runs=false
 javafx.enabled=true
 javafx.fallback.class=com.javafx.main.NoJavaFXFallback
 # Main class for JavaFX
-javafx.main.class=metrosimu.Kayttoliittyma
+javafx.main.class=Paanakyma.Main
 javafx.preloader.class=
 # This project does not use Preloader
 javafx.preloader.enabled=false
