+++ src/Paanakyma/Main.java
@@ -1,0 +1,25 @@
+
+package Paanakyma;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+public class Main extends Application{
+    private static PaanakymanKasittelija paanakymakasittelija;
+    public static void main(String[] args) {
+        launch(args);
+        
+    }
+
+    @Override
+    public void start(Stage primaryStage) throws Exception {
+        Parent root = FXMLLoader.load(getClass().getResource("Paanakyma.fxml"));
+        primaryStage.setTitle("Resurssienkeruusimulaattori");
+        primaryStage.setScene(new Scene(root));
+        primaryStage.show();
+    }
+    
+}
